fixed filesystem headers
note to self, never use * in .pro files

+++ headers/games/machineplayer.h
@@ -1,7 +1,7 @@
 #ifndef MACHINEPLAYER_H
 #define MACHINEPLAYER_H
 
-#include "player.h"
+#include "headers/games/player.h"
 
 class MachinePlayer: public Player
 {
@@ -13,14 +13,18 @@ public:
     bool will_continue();
 
 private:
-    std::map<Card, double> card_scores;
+    std::map<Card, int> card_scores;
     void update_score_map();
+
+
 };
 
 int holeDistance(Card begin, Card end, int rank_order[], const Deck &pure_suit_deck);
 int highestRank(std::vector<Card> const &cards);
 int lowestRank(std::vector<Card> const &cards);
 bool isBetween(int n, int a, int b);
-// Finds what cards can be currently placed
+
+
+
 
 #endif // MACHINEPLAYER_H
